+++ Kernel/Storage/ATA/AHCIPortHandler.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <Kernel/CommandLine.h>
-#include <Kernel/Storage/AHCIPortHandler.h>
+#include <Kernel/Storage/ATA/AHCIPortHandler.h>
 
 namespace Kernel {
