+++ Kernel/Storage/ATA/AHCIPortHandler.h
@@ -15,8 +15,8 @@
 #include <Kernel/PhysicalAddress.h>
 #include <Kernel/Random.h>
 #include <Kernel/Sections.h>
-#include <Kernel/Storage/AHCIController.h>
-#include <Kernel/Storage/AHCIPort.h>
+#include <Kernel/Storage/ATA/AHCIController.h>
+#include <Kernel/Storage/ATA/AHCIPort.h>
 #include <Kernel/Storage/StorageDevice.h>
 #include <Kernel/WaitQueue.h>
