+++ Kernel/Storage/ATA/BMIDEChannel.cpp
@@ -6,9 +6,9 @@
 
 #include <Kernel/Bus/PCI/API.h>
 #include <Kernel/Sections.h>
-#include <Kernel/Storage/ATA.h>
-#include <Kernel/Storage/BMIDEChannel.h>
-#include <Kernel/Storage/IDEController.h>
+#include <Kernel/Storage/ATA/ATA.h>
+#include <Kernel/Storage/ATA/BMIDEChannel.h>
+#include <Kernel/Storage/ATA/IDEController.h>
 #include <Kernel/WorkQueue.h>
 
 namespace Kernel {
